# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: stylifier-api
host: mock.stylifier.com
basePath: /
schemes:
  - https
consumes:
  - application/json; charset=utf-8
  - application/json
paths:
...
  '/color_pallets/{palletId}':
    x-swagger-router-controller: colorPallets
    post:
      description: bookmark a color pallet
      operationId: createColorPalletBookmart
      parameters:
        - name: palletId
          in: path
          required: true
          type: string
        - name: title
          in: query
          type: string

      responses:
        "200":
          description: color has bookmarked

    delete:
      description: delete bookmarked a color pallet
      operationId: deleteColorPalletBookmart
      parameters:
        - name: palletId
          in: path
          required: true
          type: string

      responses:
        "200":
          description: bookmarked pallet deleted
  '/color_pallets_bookmarks':
    x-swagger-router-controller: colorPallets
    get:
      description: get users color pallet bookmart
      operationId: getUsersColorPalletBookmart
      parameters:
        - name: pagination
          in: query
          type: number

      responses:
        "200":
          description: color has bookmarked
  '/color_pallets':
    x-swagger-router-controller: colorPallets
    post:
      description: register a color pallet
      operationId: createColorPallet
      parameters:
        - name: colorPallet
          in: body
          required: true
          schema:
            $ref: "#/definitions/colorPallet"

      responses:
        "200":
          description: color pallet created
          examples:
            application/json; charset=utf-8:
              {
                "success": true
              }

    get:
      description: get color pallet suggestion
      operationId: getCollorPalletSuggestion
      parameters:
        - name: color_code
          in: query
          required: true
          type: string
          pattern: '^#(?:[0-9a-fA-F]{3}){1,2}$'

      responses:
        "200":
          description: color pallet suggestions
          schema:
            $ref: '#/definitions/colorPallets'
  '/color_codes':
    x-swagger-router-controller: colorPallets
    get:
      description: get color code matching
      operationId: getColorCodes

      responses:
        "200":
          description: array of color codes

    post:
      description: add color code
      operationId: addColorCodes

      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              code:
                type: string
              name:
                type: string

      responses:
        "200":
          description: color code created
...
  '/users/{username}/products':
    x-swagger-router-controller: products
    get:
      description: get top 10 products
      operationId: getUserProducts
      parameters:
        - name: username
          in: path
          required: true
          type: string
        - name: pagination
          in: query
          type: number
        - name: name
          in: query
          required: false
          type: string
        - name: brand
          in: query
          required: false
          type: string
        - name: color
          in: query
          required: false
          type: string
        - name: sub_color
          in: query
          required: false
          type: string
        - name: category
          in: query
          required: false
          type: string
        - name: hex
          in: query
          required: false
          type: string
          pattern: '^#[0-9a-f]{3,6}$'
      responses:
        "200":
          description: array of products
          schema:
            $ref: '#/definitions/products'
...
  '/products':
    x-swagger-router-controller: products
    get:
      description: get top 10 products
      operationId: getSelfProducts
      parameters:
        - name: pagination
          in: query
          type: number
        - name: name
          in: query
          required: false
          type: string
        - name: brand
          in: query
          required: false
          type: string
        - name: color
          in: query
          required: false
          type: string
        - name: sub_color
          in: query
          required: false
          type: string
        - name: category
          in: query
          required: false
          type: string
        - name: hex
          in: query
          required: false
          type: string
          pattern: '^#[0-9a-f]{3,6}$'
      responses:
        "200":
          description: array of products
          schema:
            $ref: '#/definitions/products'

    post:
      description: create a new products
      operationId: createProduct
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/product'

      responses:
        "200":
          description: product created
          examples:
            application/json; charset=utf-8:
              {
                "success": true
              }
  '/product/{productId}/bookmark':
    x-swagger-router-controller: products
    post:
      description: bookmark a product
      operationId: createProductBookmart
      parameters:
        - name: productId
          in: path
          required: true
          type: string
        - name: pallet_id
          in: query
          required: false
          type: string
        - name: title
          in: query
          type: string

      responses:
        "200":
          description: product has bookmarked

    delete:
      description: delete bookmarked a product
      operationId: deleteProductBookmart
      parameters:
        - name: productId
          in: path
          required: true
          type: string
        - name: pallet_id
          in: query
          required: false
          type: string

      responses:
        "200":
          description: bookmarked product deleted
  '/product_bookmarks':
    x-swagger-router-controller: products
    get:
      description: get users product bookmart
      operationId: getUsersProductBookmart
      parameters:
        - name: pagination
          in: query
          type: number

      responses:
        "200":
          description: color has bookmarked
...
definitions:
  id:
    type: string
...
  style:
    type: string
  user_short:
    type: object
    properties:
      id:
        $ref: '#/definitions/id'
      username:
        type: string
      full_name:
        type: string
      profile_picture:
        type: string
  user:
    allOf:
      - $ref: '#/definitions/user_short'
      - type: object
        properties:
          bio:
            type: string
          counts:
            type: object
            properties:
              follows:
                type: number
              followed_by:
                type: number
              media:
                type: number
              thread_asked_by:
                type: number
              thread_asked_by_open:
                type: number
              sponsored:
                type: number
          contribution_earned:
            type: number
          rating:
            type: number
          request_rating:
            type: number
          is_instagram_user:
            type: boolean
          is_brand:
            type: boolean
          styles:
            type: array
            items:
              $ref: '#/definitions/style'
          sponsors:
            type: array
            items:
              $ref: '#/definitions/user_short'
          created_time:
            type: string
...
  product:
    type: object
    properties:
      name:
        type: string
      code:
        type: string
      price:
        type: object
        properties:
          original:
            type: string
          promotional:
            type: string
      user:
        $ref: '#/definitions/user'
      media:
        type: array
        items:
          $ref: '#/definitions/images'
      sizes:
        type: string
      externalURL:
        type: string
      brand:
        type: string
      category:
        type: string
      color:
        type: string
      subColor:
        type: string
      colorPallet:
        type: string
      created_time:
        type: string
  products:
    type: array
    items:
      $ref: '#/definitions/product'
...
  image:
    type: object
    properties:
      width:
        type: number
      height:
        type: number
      url:
        type: string
  images:
    type: object
    properties:
      thumbnail:
        $ref: '#/definitions/image'
      low_resolution:
        $ref: '#/definitions/image'
      standard_resolution:
        $ref: '#/definitions/image'
...
  colorPallet:
    type: object
    properties:
      id:
        type: string
      code:
        type: string
      likes:
        type: string
      popularity:
        type: number
  colorPallets:
    type: array
    items:
      $ref: '#/definitions/colorPallet'